+++ packages/flutter_bloc/lib/flutter_bloc.dart
@@ -2,6 +2,8 @@ library flutter_bloc;
 
 export './src/bloc_builder.dart';
 export './src/bloc_listener.dart';
-export './src/bloc_listener_tree.dart';
 export './src/bloc_provider.dart';
-export './src/bloc_provider_tree.dart';
+export './src/multi_bloc_listener.dart';
+export './src/multi_bloc_provider.dart';
+export './src/multi_repository_provider.dart';
+export './src/repository_provider.dart';