+++ packages/flutter_bloc/lib/src/bloc_builder.dart
@@ -11,14 +11,9 @@ typedef BlocWidgetBuilder<S> = Widget Function(BuildContext context, S state);
 
 /// Signature for the condition function which takes the previous state and the current state
 /// and is responsible for returning a `bool` which determines whether or not to rebuild
-/// `BlocBuilder` with the current state.
+/// [BlocBuilder] with the current state.
 typedef BlocBuilderCondition<S> = bool Function(S previous, S current);
 
-/// A Flutter widget which requires a [Bloc] and a [BlocWidgetBuilder] `builder` function.
-/// [BlocBuilder] handles building the widget in response to new states.
-/// BlocBuilder analogous to [StreamBuilder] but has simplified API
-/// to reduce the amount of boilerplate code needed
-/// as well as bloc-specific performance improvements.
 class BlocBuilder<E, S> extends BlocBuilderBase<E, S> {
   /// The [Bloc] that the [BlocBuilder] will interact with.
   final Bloc<E, S> bloc;
@@ -32,10 +27,35 @@ class BlocBuilder<E, S> extends BlocBuilderBase<E, S> {
   /// The `condition` function will be invoked on each bloc state change.
   /// The `condition` takes the previous state and current state and must return a `bool`
   /// which determines whether or not the `builder` function will be invoked.
-  /// The previous state will be initialized to `currentState` when the `BlocBuilder` is initialized.
+  /// The previous state will be initialized to `currentState` when the [BlocBuilder] is initialized.
   /// `condition` is optional and if it isn't implemented, it will default to return `true`.
+  ///
+  /// ```dart
+  /// BlocBuilder(
+  ///   bloc: BlocProvider.of<BlocA>(context),
+  ///   condition: (previousState, currentState) {
+  ///     // return true/false to determine whether or not
+  ///     // to rebuild the widget with currentState
+  ///   },
+  ///   builder: (context, state) {
+  ///     // return widget here based on BlocA's state
+  ///   }
+  ///)
+  /// ```
   final BlocBuilderCondition<S> condition;
 
+  /// [BlocBuilder] handles building a widget in response to new states.
+  /// [BlocBuilder] is analogous to [StreamBuilder] but has simplified API
+  /// to reduce the amount of boilerplate code needed as well as bloc-specific performance improvements.
+  ///
+  /// ```dart
+  /// BlocBuilder(
+  ///   bloc: BlocProvider.of<BlocA>(context),
+  ///   builder: (BuildContext context, BlocAState state) {
+  ///   // return widget here based on BlocA's state
+  ///   }
+  /// )
+  /// ```
   const BlocBuilder({
     Key key,
     @required this.bloc,
@@ -49,13 +69,13 @@ class BlocBuilder<E, S> extends BlocBuilderBase<E, S> {
   Widget build(BuildContext context, S state) => builder(context, state);
 }
 
-/// Base class for widgets that build themselves based on interaction with
-/// a specified [Bloc].
-///
-/// A [BlocBuilderBase] is stateful and maintains the state of the interaction
-/// so far. The type of the state and how it is updated with each interaction
-/// is defined by sub-classes.
 abstract class BlocBuilderBase<E, S> extends StatefulWidget {
+  /// Base class for widgets that build themselves based on interaction with
+  /// a specified [Bloc].
+  ///
+  /// A [BlocBuilderBase] is stateful and maintains the state of the interaction
+  /// so far. The type of the state and how it is updated with each interaction
+  /// is defined by sub-classes.
   const BlocBuilderBase({Key key, this.bloc, this.condition}) : super(key: key);
 
   /// The [Bloc] that the [BlocBuilderBase] will interact with.