+++ packages/flutter_bloc/CHANGELOG.md
@@ -1,6 +1,45 @@
+# 0.19.0
+
+Addresses [#354](https://github.com/felangel/bloc/issues/354)
+
+#### BlocProvider
+
+- Refactor `BlocProvider` to extend `Provider`
+- Rename `BlocProviderTree` to `MultiBlocProvider`
+
+#### ImmutableProvider
+
+- Refactor `ImmutableProvider` to extend `Provider`
+- Rename `ImmutableProvider` to `RepositoryProvider`
+- Rename `ImmutableProviderTree` to `MultiRepositoryProvider`
+
+#### BlocListener
+
+- Rename `BlocListenerTree` to `MultiBlocListener`
+
+#### Documentation
+
+- Inline documentation updates/improvements
+
+# 0.18.3
+
+Fix `BlocProvider` bug where `copyWith` does not preserve `dispose` value ([#376](https://github.com/felangel/bloc/issues/376)).
+
+# 0.18.2
+
+Fix `BlocListener` bug where `listener` gets called even when no state change occurs ([#368](https://github.com/felangel/bloc/issues/368)).
+
+# 0.18.1
+
+Minor Documentation Updates
+
+# 0.18.0
+
+Expose `ImmutableProvider` & `ImmutableProviderTree` to enable developers to provide immutable values, such as repositories, throughout the widget tree ([#364](https://github.com/felangel/bloc/pull/364)) and Documentation Updates
+
 # 0.17.0
 
-Update `BlocProvider` to automatically `dispose` the provided bloc ([#349](https://github.com/felangel/bloc/pull/349)) and Documentation Updates.
+Update `BlocProvider` to automatically `dispose` the provided bloc ([#349](https://github.com/felangel/bloc/pull/349)) and Documentation Updates
 
 # 0.16.0